requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=1.0.0
openai>=1.3.0
PyPDF2>=3.0.0
faiss-cpu>=1.7.4
tiktoken>=0.5.1
streamlit>=1.22.0
PyMuPDF>=1.21.1 